// src/frontend/src/api/entities/partner.entity.ts
export interface PartnerEntity {
   readonly id: string;
   readonly partnerName: string;
   readonly sector: string;
   readonly contactEmail: string;
   readonly contactPhone: string;
   readonly rateForProject: number;
   readonly branch: string;
   readonly projectName: string;
   readonly organizationType: string;
   readonly description: string;
   readonly representativeJob: string;
   readonly isActive: boolean;
   readonly createdAt: string;
   readonly updatedAt: string;
   
}